Move Page-related endpoints to its own router
Not only these functions are logically different that browser-related
endpoints, this will also make it easier to add the distillation
functionality on the page.
